requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
firebase-admin>=6.4.0
jinja2==3.1.3
feedparser==6.0.11
python-dateutil==2.8.2
httpx==0.26.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
openai>=1.55.0
apscheduler==3.10.4
